+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 2020-3-10 1.10.09-15
+# 2020-3-10 1.10.09-18
 - Search added to Playlist Edit->Add Tracks
 - Searching in All Tracks includes Artist name value
 - Editing playlist shows rho start/stop values as it would be played, with connection lines marked
@@ -16,9 +16,10 @@
 - Versioning in 7 tap updated
 - Find More Tracks removed from UI
 - Show created_by_name credit on Track lists
+- Delay before restarting polling on Reset/Restart/Reboot (and UI without Rescan buttons during that time)
 # 2020-3-12 1.10.08 1d03f3071c64485f403b5ded6412401369d7fa77
 - Fetching more artist spinner fix
-# 2020-3-4 1.10.00-05
+# 2020-3-4 1.10.00-05 e1da68217ebdb8fa0610bdc797c7eb3e40b28345
 - wording and UI changes
 - Description error checking
 - Community log-in / sign-up error checking updated.